feat: Position의 equals, hashCode 오버라이딩
